revert: "chore: Separate server start and api setup logic"
This reverts commit 6e03171ae34d90f5b62c2ecca6b899b8e2053696.
